refactor complete task functions into Task protocol
