+++ lib/mozart/task/receive.ex
@@ -15,7 +15,3 @@ defmodule Mozart.Task.Receive do
     type: :receive
   ]
 end
-
-defimpl Mozart.Task, for: Mozart.Task.Receive do
-  def completable(receive), do: receive.complete
-end
